- These changes enable consistent validation results between pruned and archival nodes and accelerate the hash to leaf index queries. - Modified the blockchain db is_stxo function to rather query the deletion status of the UTXO MMR instead of the blockchain backend stxo_db. This change is required to allow pruned nodes and archival nodes to have the same validation results when checking if new outputs were previous STXOs. The store STXOs of a pruned node is discarded when it enters the horizon state. - Added fetch_mmr_leaf_index to the blockchain backend trait and provided implementations for the memory and lmdb backends. This function is used to implement the is_stxo logic in blockchain db. - Finding the leaf index from a hash was a slow operation, this was accelerated by adding a position function to the ArrayLike trait and providing an optimised implementation for the MemDbVec that is used to store the hashes in the MMR. Inefficient versions were added to the other ArrayLike implementations for completeness but these won't be used. - The MemDbVec was accelerated by making use of two hashmaps to allow quick mapping from index to item and item to index. This will accelerate the fetch_mmr_leaf_index function allowing the is_stxo checks to be performed efficiently. - The find_leaf_index function of the MerkleMountainRange was updated to make use of the new position function of the ArrayLike trait and find_leaf_index was added to the MmrCache to allow both the current and base Mmr to be checked. - Added the pruned_mode_is_stxo test to check the pruned mode is_stxo functionality and updated the merging_and_fetch_checkpoints_and_stxo_discard test. - Added the find_leaf_index test to check the MerkleMountainRange changes.

+++ base_layer/mmr/src/mutable_mmr.rs
@@ -133,7 +133,7 @@ where
     }
 
     /// See [MerkleMountainRange::find_leaf_index]
-    pub fn find_leaf_index(&self, hash: &Hash) -> Result<Option<usize>, MerkleMountainRangeError> {
+    pub fn find_leaf_index(&self, hash: &Hash) -> Result<Option<u32>, MerkleMountainRangeError> {
         self.mmr.find_leaf_index(hash)
     }
